fix(modal): mark for check after update of modal options
fix #4801
