+++ src/modal/modal-ref.ts
@@ -5,8 +5,7 @@ import { takeUntil } from 'rxjs/operators';
 
 import { NgbModalBackdrop } from './modal-backdrop';
 import { NgbModalWindow } from './modal-window';
-import { NgbModalOptions, NgbModalUpdatableOptions } from './modal-config';
-import { isDefined } from '../util/util';
+import { NgbModalUpdatableOptions } from './modal-config';
 
 import { ContentRef } from '../util/popup';
 import { isPromise } from '../util/util';
@@ -39,22 +38,6 @@ export class NgbActiveModal {
 	dismiss(reason?: any): void {}
 }
 
-const WINDOW_ATTRIBUTES: string[] = [
-	'animation',
-	'ariaLabelledBy',
-	'ariaDescribedBy',
-	'backdrop',
-	'centered',
-	'fullscreen',
-	'keyboard',
-	'role',
-	'scrollable',
-	'size',
-	'windowClass',
-	'modalDialogClass',
-];
-const BACKDROP_ATTRIBUTES: string[] = ['animation', 'backdropClass'];
-
 /**
  * A reference to the newly opened modal returned by the `NgbModal.open()` method.
  */
@@ -65,31 +48,15 @@ export class NgbModalRef {
 	private _resolve: (result?: any) => void;
 	private _reject: (reason?: any) => void;
 
-	private _applyWindowOptions(windowInstance: NgbModalWindow, options: NgbModalOptions): void {
-		WINDOW_ATTRIBUTES.forEach((optionName: string) => {
-			if (isDefined(options[optionName])) {
-				windowInstance[optionName] = options[optionName];
-			}
-		});
-	}
-
-	private _applyBackdropOptions(backdropInstance: NgbModalBackdrop, options: NgbModalOptions): void {
-		BACKDROP_ATTRIBUTES.forEach((optionName: string) => {
-			if (isDefined(options[optionName])) {
-				backdropInstance[optionName] = options[optionName];
-			}
-		});
-	}
-
 	/**
 	 * Updates options of an opened modal.
 	 *
 	 * @since 14.2.0
 	 */
 	update(options: NgbModalUpdatableOptions): void {
-		this._applyWindowOptions(this._windowCmptRef.instance, options);
+		this._windowCmptRef.instance.updateOptions(options);
 		if (this._backdropCmptRef && this._backdropCmptRef.instance) {
-			this._applyBackdropOptions(this._backdropCmptRef.instance, options);
+			this._backdropCmptRef.instance.updateOptions(options);
 		}
 	}
